+++ src/Wolverine/TestMessageContext.cs
@@ -237,7 +237,7 @@ public class TestMessageContext : IMessageContext
         return EmptyAsyncEnumerable<TResponse>(cancellation);
     }
 
-    Task<TResponse> ICommandBus.InvokeStreamAsync<TRequest, TResponse>(IAsyncEnumerable<TRequest> messages,
+    Task<TResponse> ICommandBus.StreamAsync<TRequest, TResponse>(IAsyncEnumerable<TRequest> messages,
         CancellationToken cancellation, TimeSpan? timeout)
     {
         _invoked.Add(messages);
@@ -246,7 +246,7 @@ public class TestMessageContext : IMessageContext
         return Task.FromResult(response);
     }
 
-    Task<TResponse> ICommandBus.InvokeStreamAsync<TRequest, TResponse>(IAsyncEnumerable<TRequest> messages,
+    Task<TResponse> ICommandBus.StreamAsync<TRequest, TResponse>(IAsyncEnumerable<TRequest> messages,
         DeliveryOptions options, CancellationToken cancellation, TimeSpan? timeout)
     {
         var envelope = new Envelope(messages);